feat(VideoModal): Move modal to commons
