+++ src/commons/VideoGallery/Gallery/Category.jsx
@@ -1,7 +1,7 @@
 // @flow
 import React from 'react';
 import { Card } from '@windingtree/wt-ui-react';
-import VideoModal from './VideoModal';
+import VideoModal from 'commons/VideoModal';
 import './styles.scss';
 
 type PropsType = {
