+++ src/commons/VideoModal/index.jsx
@@ -1,0 +1,60 @@
+// @flow
+import React from 'react';
+import YouTube from 'react-youtube';
+import { Modal } from '@windingtree/wt-ui-react';
+
+type PropsType = {
+  onHide?: (SyntheticMouseEvent<*>) => void,
+  show: boolean,
+  title: string,
+  videoSrc: string
+};
+
+
+type StateType = {
+  showModal: boolean
+};
+
+class VideoModal extends React.Component<PropsType, StateType> {
+  state = { showModal: false };
+
+  hanldeOnClick = (e: SyntheticMouseEvent<*>) => {
+    e.preventDefault();
+    this.setState({ showModal: true });
+  }
+
+  hideModal = () => {
+    this.setState({ showModal: false });
+  };
+
+  render() {
+    const {
+      show, title, videoSrc, onHide, ...otherProps
+    } = this.props;
+    const opts = {
+      playerVars: {
+        autoplay: 1,
+        accelerometer: 1,
+        'encrypted-media': 1,
+        gyroscope: 1,
+        'picture-in-picture': 1,
+      },
+    };
+    return (
+      <Modal {...otherProps} show={show} className="modal-media" onHide={onHide}>
+        <Modal.Body>
+          <div className="video-fluid">
+            <YouTube
+              videoId={videoSrc}
+              className="youtube-video"
+              opts={opts}
+              onHide={onHide}
+            />
+          </div>
+        </Modal.Body>
+      </Modal>
+    );
+  }
+}
+
+export default VideoModal;
